КЛиенты. Импорт csv

+++ app/Http/Controllers/Operator/RoleController.php
@@ -7,22 +7,21 @@ use Illuminate\Http\Request;
 
 class RoleController extends BaseController {
 
-   public function index(Request $request)
-{
-    $sort = $request->get('sort', 'id');
-    $direction = $request->get('direction', 'asc');
-    $search = $request->get('name', '');
+    public function index(Request $request) {
+        $sort = $request->get('sort', 'id');
+        $direction = $request->get('direction', 'asc');
+        $search = $request->get('name', '');
 
-    $rolesQuery = Role::query();
+        $rolesQuery = Role::query();
 
-    if ($search) {
-        $rolesQuery->where('name', 'like', "%$search%");
+        if ($search) {
+            $rolesQuery->where('name', 'like', "%$search%");
+        }
+
+        $roles = $rolesQuery->orderBy($sort, $direction)->paginate(10);
+
+        return view('roles.index', compact('roles', 'sort', 'direction', 'search'));
     }
-
-    $roles = $rolesQuery->orderBy($sort, $direction)->paginate(10);
-
-    return view('roles.index', compact('roles', 'sort', 'direction', 'search'));
-}
 
     public function create() {
         return view('roles.create');
@@ -57,8 +56,16 @@ class RoleController extends BaseController {
     }
 
     public function destroy(Role $role) {
+        // Проверка, есть ли пользователи с этой ролью
+        if ($role->users()->exists()) {
+            return redirect()->route('roles.index')
+                            ->with('error', "Невозможно удалить роль, так как она используется у user");
+        }
+
         $role->delete();
-        return redirect()->route('roles.index')->with('success', 'Роль удалена.');
+
+        return redirect()->route('roles.index')
+                        ->with('success', 'Роль удалена.');
     }
 
 }
